docker/setup/.bin/PluginInstaller/requirements.txt
loguru
pyyaml
requests
